refactor(get-user): fix shop rating type

+++ internal/domain/shop/model/shop.go
@@ -9,7 +9,7 @@ import (
 type Shop struct {
 	ID         int       `json:"id"`
 	Name       string    `json:"name"`
-	Rating     float32   `json:"rating"`
+	Rating     float64   `json:"rating"`
 	JoinedDate time.Time `json:"joinedDate"`
 	UserID     int       `json:"userId"`
